Access parameters from a configuration file

Access parameters from a configuration file

FINANCIAL.TIMESTAMP(file = get_config())

timestamps_datastore_export(file = get_config())

FINANCIAL.TIMESTAMP(file = get_config())

ALD.TIMESTAMP(file = get_config())

DATASTORE.TIMESTAMP(file = get_config())

DATAPREP.TIMESTAMP(file = get_config())

START.YEAR(file = get_config())

TIME.HORIZON(file = get_config())

RISK.YEAR(file = get_config())

ADDITIONAL.YEAR(file = get_config())

TECH.LIST(file = get_config())

TECH.EXCLUDE(file = get_config())

SECTOR.LIST(file = get_config())

OTHER.SECTOR.LIST(file = get_config())

SCENARIO.SOURCES.LIST(file = get_config())

IEA.SCENARIO.LIST(file = get_config())

WEB.REGION.LIST(file = get_config())

SCENARIO.GEOGRAPHIES.LIST(file = get_config())

EQUITY.MARKET.LIST(file = get_config())

GLOBAL.AGGREGATE.SECTOR.LIST(file = get_config())

GLOBAL.AGGREGATE.SCENARIO.SOURCES.LIST(file = get_config())

CREATE.META.PORTFOLIO(file = get_config())

META.INVESTOR.NAME(file = get_config())

META.PORTFOLIO.NAME(file = get_config())

HasBV(file = get_config())

HasRISK(file = get_config())

HasMAP(file = get_config())

HasSB(file = get_config())

inc_metaportfolio(file = get_config())

inc_project_metaportfolio(file = get_config())

allowable_asset_list(file = get_config())

Arguments

file
A string giving a path to a configuration file.

Value

The value of the requested field. It's base time depends on the field.

See also

Examples

# Using the default configuration file that comes with the package
FINANCIAL.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "DONT-DELETE"

# Using your configuration file
path_to_your_config_file <- example_config("config-toy.yml")

# Passing your configuration file via the `file` argument
FINANCIAL.TIMESTAMP(file = path_to_your_config_file)

#> [1] "2018Q3"

# The default remains the same
FINANCIAL.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "DONT-DELETE"

ALD.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] 201812

# Setting the path to your configuration file globally
old_options <- options(r2dii_config = path_to_your_config_file)
# The default configuration file has now changed
FINANCIAL.TIMESTAMP()

#> [1] "2018Q3"

ALD.TIMESTAMP()

#> [1] 3333

# Reset options to cleanup
options(old_options)
timestamps_datastore_export()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "ASK-MAURO"

FINANCIAL.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "DONT-DELETE"

DATASTORE.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "ASK-MAURO"

DATAPREP.TIMESTAMP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "mauroexplores"

START.YEAR()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] 2019

TIME.HORIZON()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] 5

RISK.YEAR()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] 2020

ADDITIONAL.YEAR()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] 2040

TECH.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#>  [1] "Electric"      "Hybrid"        "ICE"           "GasCap"       
#>  [5] "CoalCap"       "OilCap"        "RenewablesCap" "HydroCap"     
#>  [9] "NuclearCap"    "Coal"          "Oil"           "Gas"          

TECH.EXCLUDE()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#>  [1] "FuelCell"                           "Fuel cell"                         
#>  [3] "OtherCap"                           "OtherFF"                           
#>  [5] "Bof Shop"                           "Coking Plant"                      
#>  [7] "Sintering Plant"                    "Direct Or Smelting Reduction Plant"
#>  [9] "Pelletizing Plant"                  "Open Hearth Meltshop"              

SECTOR.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "Automotive"   "Power"        "Fossil Fuels" "Oil&Gas"      "Coal"        

OTHER.SECTOR.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "Shipping" "Steel"    "Aviation" "Cement"  

SCENARIO.SOURCES.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "GP_ER_2015" "ETP2017"    "BNEF2017"   "WEO2018"    "SBTI"      

IEA.SCENARIO.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> NULL

WEB.REGION.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "Global"           "OECD"             "NonOECD"          "NorthAmerica"    
#> [5] "Europe"           "AsiaPacific"      "OECDNorthAmerica"

SCENARIO.GEOGRAPHIES.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "GlobalAggregate" "Global"          "NonOECD"         "OECD"           
#> [5] "Europe"          "NorthAmerica"   

EQUITY.MARKET.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "Global"

GLOBAL.AGGREGATE.SECTOR.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "Power"

GLOBAL.AGGREGATE.SCENARIO.SOURCES.LIST()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] "WEO2018" "ETP2017"

CREATE.META.PORTFOLIO()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> NULL

META.INVESTOR.NAME()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> NULL

META.PORTFOLIO.NAME()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> NULL

HasBV()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] FALSE

HasRISK()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] TRUE

HasMAP()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] TRUE

HasSB()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

#> [1] FALSE

config <- create_config("
  default:
    ComparisonBenchmarks:
      CreateMetaPortfolio: a_value
")

#> ✔ Saved to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpU1C1YZ/config-a4d54d3ddb6.yml'.

inc_metaportfolio(config)

#> [1] "a_value"

config <- create_config("
  default:
    ComparisonBenchmarks:
      CreateProjectMetaPortfolio: a_value
")

#> ✔ Saved to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpU1C1YZ/config-a4d2181033c.yml'.

inc_project_metaportfolio(config)

#> [1] "a_value"

config <- create_config("
  default:
    Lists:
      AssetTypes: a_value
")

#> ✔ Saved to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpU1C1YZ/config-a4d1cfb6953.yml'.

allowable_asset_list(config)

#> [1] "a_value"