Create functions to access configuration parameters

This is a function factory to create functions which output the value of a configuration parameter.

get_param(..., if_null = NULL)

Arguments

...
Character string(s) giving a(the) field(s) in a configuration file.

if_null
A value to fill an unset parameter, or the function stop.

Value

A function.

See also

Examples

# get_param() is a function factory
# Create a `get_a_value()` function
get_a_value <- get_param("a_field")
class(get_a_value)

#> [1] "function"

# Create an example configuration file
config <- glue::glue("
    default:
      a_field: a_value
  ")
show_config(config)

#>   default:
#>     a_field: a_value

# Write the config file to a temporary file
path_to_config_file <- tempfile()
readr::write_lines(config, path_to_config_file)

# Use `get_a_value()` to get a value from a field in the example config file
get_a_value(file = path_to_config_file)

#> [1] "a_value"

# Fails: The parameter 'bad' doesn't exist, so it is `NULL`
fails <- get_param("bad", if_null = stop)
try(fails())

#> Error : bad must be not `NULL`.
#> Please set bad on config.yml.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(fails())
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─r2dii.utils:::fails()
#>  40.                             └─r2dii.utils:::pluck_param(dots, file = file, if_null = if_null)

returns_null <- get_param("bad")
returns_null()

#> NULL

returns_value_passed_to_if_null <- get_param("bad", if_null = 1)
returns_value_passed_to_if_null()

#> Warning: On config.yml, bad is undefined.

#> Setting bad to default value: 1.

#> [1] 1