Paths to example data installed in any package at .libPaths()

Paths to example data installed in any package at .libPaths()

path_examples(package)

path_example(path, package)

Arguments

package
Character string giving the name of a single package. NULL defaults to the current

path
Name of file.

Value

path_example(): A path to a single example.

path_examples(): A character vector of all built-in examples.

See also

Examples

path_examples("readr")

#> [1] "challenge.csv"     "epa78.txt"         "example.log"      
#> [4] "fwf-sample.txt"    "massey-rating.txt" "mtcars.csv"       
#> [7] "mtcars.csv.bz2"    "mtcars.csv.zip"   

path_example("mtcars.csv", package = "readr")

#> [1] "/Users/runner/runners/2.169.1/work/_temp/Library/readr/extdata/mtcars.csv"