Show a configuration file

Show a configuration file

show_config(path = get_config(), ...)

Arguments

path
A string giving the path to a configuration file. NULL defaults to get_config().

...
Other arguments passed to readr::read_lines(), e.g. n_max.

Value

Prints to the console the configuration file.

See also

Examples

# A configuration file stored in the package
path <- default_config()

#> Warning: Using a toy configuration file.
#> You may use your configuration file with `options(r2dii_config = 'path/to/config.yml')`

show_config(path, n_max = 10)

#> default:
#>   Methodology:
#>     HasBookValue: FALSE # ignore
#>     HasRISK: TRUE # ignore a flag specifying if the physical risk assessment should be done
#>     HasMAP: TRUE # flag to specify if map data (production per country) should be prepared
#>     HasSB: FALSE # ignore for now
#> 
#>   TimeStamps:
#>     # setting the dates of the data-input for financial and asset-level data: this identifies the input files
#>     # The name "DONT-DELETE" is to alert people to not delete this directory

# Create a new configuration file to show a more general case
config_file <- glue::glue("
  default:
    a_field: a_value
")
config_path <- tempfile()
readr::write_lines(config_file, config_path)

show_config(config_path)

#>   default:
#>     a_field: a_value